Support "[role=button]" style as well
Material UI has a tendency to use divs and spans, instead of proper semantic html for some reason. But they are careful to add role=button to things that should act as buttons.
Since this is in the spirit of what the selector wants to target, maybe its worth adding it here?